# internal/pb/api/v1/drones/drones.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/drones/drones.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Drones"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/drones": {
      "post": {
        "operationId": "Drones_CreateDrone",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dronesCreateDroneResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dronesCreateDroneRequest"
            }
          }
        ],
        "tags": [
          "Drones"
        ]
      }
    },
    "/v1/drones/authors/{login}": {
      "get": {
        "operationId": "Drones_GetDronesByAuthor",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dronesGetDronesByAuthorResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Drones"
        ]
      }
    },
    "/v1/drones/{id}": {
      "get": {
        "operationId": "Drones_GetDroneByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dronesGetDroneByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Drones"
        ]
      },
      "delete": {
        "operationId": "Drones_DeleteDrone",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Drones"
        ]
      },
      "put": {
        "operationId": "Drones_UpdateDrone",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dronesUpdateDroneResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dronesDronesUpdateDroneBody"
            }
          }
        ],
        "tags": [
          "Drones"
        ]
      }
    },
    "/v1/drones/{id}/mission": {
      "post": {
        "operationId": "Drones_StartDroneMission",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dronesDronesStartDroneMissionBody"
            }
          }
        ],
        "tags": [
          "Drones"
        ]
      }
    }
  },
  "definitions": {
    "dronesCreateDroneRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        }
      }
    },
    "dronesCreateDroneResponse": {
      "type": "object",
      "properties": {
        "drone": {
          "$ref": "#/definitions/modelDrone"
        }
      }
    },
    "dronesDronesStartDroneMissionBody": {
      "type": "object",
      "properties": {
        "mission": {
          "$ref": "#/definitions/modelMission"
        }
      }
    },
    "dronesDronesUpdateDroneBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/modelDroneStatus"
        }
      }
    },
    "dronesGetDroneByIDResponse": {
      "type": "object",
      "properties": {
        "drone": {
          "$ref": "#/definitions/modelDrone"
        }
      }
    },
    "dronesGetDronesByAuthorResponse": {
      "type": "object",
      "properties": {
        "drones": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/modelDrone"
          }
        }
      }
    },
    "dronesUpdateDroneResponse": {
      "type": "object",
      "properties": {
        "drone": {
          "$ref": "#/definitions/modelDrone"
        }
      }
    },
    "modelCoordinate": {
      "type": "object",
      "properties": {
        "latitude": {
          "type": "number",
          "format": "float"
        },
        "longitude": {
          "type": "number",
          "format": "float"
        }
      }
    },
    "modelDrone": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/modelDroneStatus"
        },
        "createdBy": {
          "type": "string"
        }
      }
    },
    "modelDroneStatus": {
      "type": "string",
      "enum": [
        "DRONE_STATUS_AVAILABLE",
        "DRONE_STATUS_IN_MISSION",
        "DRONE_STATUS_CHARGING",
        "DRONE_STATUS_MAINTENANCE",
        "DRONE_STATUS_OFFLINE"
      ],
      "default": "DRONE_STATUS_AVAILABLE"
    },
    "modelMission": {
      "type": "object",
      "properties": {
        "coordinates": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/modelCoordinate"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}